Day 1

+++ 2025/download_input.py
@@ -24,16 +24,17 @@ current_time = datetime.datetime.now(timezone('EST'))
 day = current_time.day
 year = current_time.year
 
+
 base_dir = dirname(abspath(__file__))
 with open(join(base_dir, 'session.txt'), 'r') as sess:
     session_cookie = sess.readline().strip()
 
-
-url = f"https://adventofcode.com/{year}/day/{day}/input"
-resp = requests.get(url, cookies={'session': session_cookie})
 day_dir = join(base_dir, str(day))
 if not exists(day_dir):
     mkdir(day_dir)
 output_path = join(day_dir, 'input.txt')
-with open(output_path, 'w') as outfile:
-    outfile.write(resp.text)
+if not exists(output_path):
+    url = f"https://adventofcode.com/{year}/day/{day}/input"
+    resp = requests.get(url, cookies={'session': session_cookie})
+    with open(output_path, 'w') as outfile:
+        outfile.write(resp.text)
